1-listint_len.c

+++ 0x13-more_singly_linked_lists/1-listint_len.c
@@ -2,11 +2,11 @@
 /**
  * listint_len - returns the number of elements in a linked list
  * @h: pointer to the list
- * Return: the number number of elements
+ * Return: the number of elements
  */
 size_t listint_len(const listint_t *h)
 {
-	size_t elements = 0;
+	unsigned int elements = 0;
 
 	while (h)
 	{
